==> 4. Analysis Part 1 - tOC and mean thickness.Rmd <==
---
title: "Analysis Part 1 - tOC and mean thickness"
Authors: Serena Verdi and James Cole 
Adapted to QMIN-MC sample by candidate HZWC3
output: html_notebook
---
This script achieves the following:
(1) Draws density plots for tOC to determine which type of analyses to use - Robust linear regression and Mann-Whitney Rank Sum paired tests due to small sample sizes, heteroscedascticity, non-linearity between variables, presence of outliers and non-normal data (including non-normal residuals)
(2) Performs linear regression for Mean cortical thickness,draws box plot diagrams and performs pairwise Wilcoxon comparisons to test signficance 
(3) Investigates these questions using robust linear regression, Robust Wald test and FDR correct of p-values (when needed) to determine whether the variance of tOC can be explained by demographic factors:
  -  Question 1: Does sex have an impact on tOC, controlling for age?
  -  Question 2: Does diagnosis have an impact on tOC, controlling for age and sex?
  -  Question 3: Does sex and diagnosis combined have an impact on tOC, controlling for age? 
  -  Question 4: Do Years of education have an impact on tOC, controlling for age and sex? 
  -  Question 5: Does diagnosis and education have an impact on tOC, controlling for age and sex?
(4) Investigates whether tOC has an impact on total ACE-III and ACE-R scores as well as ACE-III and ACE-R language scores using robust linear regression, Robust Wald test and FDR correct
(5) Draws robust linear regression graphs for tOC and total and language ACE-R/ACE-III scores to view the relationships between them while accounting for outliers
---
#Load libraries 
```{r}
rm(list = ls())
library(cowplot)
library(data.table)
library(e1071)
library(dplyr)
library(ggplot2)
library(plot.matrix)
library(proxy)
library(psych)
library(tidyr)
library(stringr)
library(RPMG)
library(viridis)
library(effectsize)
library(corrplot)
library(ggseg)
library(ggsegExtra) ## you may need this for the proportional outlier maps depending on your R version
library(ggsegDesterieux) ## this is important for the proportional outlier maps 
library(ggpubr)
library(lme4)
library(lmtest)
library(arsenal)
library(lmerTest)
library(MuMIn)
library(effects)
library(ggeffects)
library(readr)
library(dplyr)
library(robustbase)


```
#Load files

```{r}
df <- read.csv("/Users/snehaprasanna/Desktop/Serena Verdi Project/QMIN/Final files/toc_df.csv")
df$diagnosis <- as.factor(df$diagnosis)
df$sex<- as.factor(df$sex)
df_ACE3_filtered <- read.csv("/Users/snehaprasanna/Desktop/Serena Verdi Project/QMIN/Final files/df_ACE3_filtered.csv")
df_ACER_filtered <- read.csv("/Users/snehaprasanna/Desktop/Serena Verdi Project/QMIN/Final files/df_ACER_filtered.csv")
df_edu_filtered <- read.csv("/Users/snehaprasanna/Desktop/Serena Verdi Project/QMIN/Final files/df_edu_filtered.csv")
```
#Plot tOC density - dependent variable
```{r}
ggplot(df, aes(x = tOC)) + 
  geom_density() + 
  ggtitle("Distribution of tOC") +
  xlab("tOC") +
  ylab("Density") +
  theme_minimal()

#Result: tOC is not normally distributed, see diagnostic scripts for the evaluation of residuals to choose between linear and robust linear regression 
```
#----------------------------------------------
# 1.	Discuss mean cortical thickness among diagnostic groups 
#----------------------------------------------
```{r}
# Descriptive statistics by group 
describeBy(df$Mean_Thickness, df$diagnosis)
d<-(describeBy(df$Mean_Thickness, df$diagnosis)) 
```
# This draws a box plot for the Mean thickness for different groups 
```{r}
df.edit <- df
df.edit$diagnosis <- gsub("Mixed dementia \\(AD and vascular dementia\\)", "Mixed dementia", df.edit$diagnosis)
df.edit$diagnosis <- gsub("Other non-AD Dementias", "Other non-AD dementias", df.edit$diagnosis)
df.edit$diagnosis <- gsub("Frontotemporal dementia", "FTD", df.edit$diagnosis)
df.edit <- df.edit %>% dplyr::rename("Diagnosis" = "diagnosis")

Mean_thickness <- ggplot(df.edit, aes(y = Mean_Thickness, x = Diagnosis, fill = Diagnosis, alpha = Diagnosis)) +
  geom_boxplot(varwidth = T, outlier.shape = 1) + 
  theme_minimal() +
  theme(text = element_text(family = "Arial", size = 14, colour = "black"),panel.grid = element_blank(),
    panel.background = element_blank(),
    axis.line = element_line(colour = "black"),
    axis.ticks = element_line(colour = "black"),
    axis.ticks.length = unit(0.2, "cm"),
    axis.text = element_text(colour = "black")  # Ensures axis numbers are black
  ) +
  labs(colour = "Diagnosis", y = "Mean Cortical Thickness", x = "Diagnosis", title = "Mean cortical thickness by diagnostic groups")  + 
  theme(text = element_text(family = "Arial", size = 14)) + 
  scale_fill_manual(name = "Diagnosis", values = c( "AD" = "grey65","MCI" = "mediumturquoise","FTD" = "lightslateblue","Mixed dementia" = "green","Other non-AD dementias" = "red")) +
  scale_alpha_manual(name = "Diagnosis", values = c(0.3, 0.3, 0.3, 0.3, 0.3))

print(Mean_thickness)

```
#Linear regression - non-adjusted model
#Why did we use linear regression? Residuals were approximately normally distributed on Q-Q plots 
```{r}
# Linear regression
summary(lm(Mean_Thickness ~ diagnosis, data = df))
s3 <- lm(Mean_Thickness ~ diagnosis, data = df)
confint(s3)
rm(s3)

 #Group by the diagnosis column and summarize the Mean_Thickness with desired statistics
df %>%
  group_by(diagnosis) %>%
  summarize(
    mean = mean(Mean_Thickness, na.rm = TRUE),
    sd = sd(Mean_Thickness, na.rm = TRUE),
    median = median(Mean_Thickness, na.rm = TRUE),
    IQR = IQR(Mean_Thickness, na.rm = TRUE),
    min = min(Mean_Thickness, na.rm = TRUE),
    max = max(Mean_Thickness, na.rm = TRUE),
    n = n()
  )

# Create the density plot for tOC by diagnosis
Mean_density <- ggplot(df, aes(x = Mean_Thickness, color = diagnosis, fill = diagnosis, name = "Diagnosis")) +
  geom_density(alpha = 0.3) +
  labs(title = "tOC density by diagnostic groups",
       x = "tOC",
       y = "Density") +
  theme_minimal() + 
  theme(
    text = element_text(family = "Arial", size = 14, colour = "black"),
    panel.grid = element_blank(),
    panel.background = element_blank(),
    axis.line = element_line(colour = "black"),
    axis.ticks = element_line(colour = "black"),
    axis.ticks.length = unit(0.2, "cm"),
    axis.text = element_text(colour = "black")) 

print(Mean_density)
```
#Linear regression - Adjusted model
#Why did we use linear regression? Residuals were approximately normally distributed on Q-Q plots 
```{r}
#Interpretation
#Model summary:
#Age and MCI diagnosis are significant predictors of Mean_Thickness.
#Other diagnoses and sex do not significantly affect Mean_Thickness.

#ANOVA:
#Both diagnosis and age are significant contributors to Mean_Thickness variance.

#Tukey’s Test:
#Mean_Thickness for MCI is significantly different from AD and Frontotemporal dementia, but other group comparisons are not significant.

summary(lm(Mean_Thickness ~ diagnosis + age + sex, data = df))
s4 <- lm(Mean_Thickness ~ diagnosis + age + sex, data = df)
confint(s4)
anova(lm(Mean_Thickness ~ diagnosis + age + sex, data = df))
rm(s4)

TukeyHSD(aov(Mean_Thickness ~ diagnosis, data = df))

```

```{r}

# Perform pairwise Mann-Whitney tests for Mean_Thickness
pairwise_comparisons_thickness <- pairwise.wilcox.test(df$Mean_Thickness, df$diagnosis, p.adjust.method = "fdr")

# Print the results
print(pairwise_comparisons_thickness)
```
#----------------------------------------------           
## 2. How does tOC vary between the groups? 
#----------------------------------------------
# Main source code on density retrieved from: http://www.sthda.com/english/wiki/ggplot2-density-plot-quick-start-guide-r-software-and-data-visualization 
# Main source code on Pairwise comparisons adapted from: ADNI scripts used within Verdi et al., 2023 paper 
```{r}
#Key Points:
#High P-Values (All Adjusted): The p-values are all quite high (many are 1.00 after Bonferroni correction), indicating that there are no significant differences in the tOC values between any pairs of diagnostic groups after adjusting for multiple comparisons.
#No Significant Differences: None of the comparisons show statistically significant differences at conventional significance levels (e.g., p < 0.05).

#Descriptive statistics ove total_outlier score
describeBy(df$tOC, df$diagnosis, IQR = T)

# Create the density plot for tOC by diagnosis
tOC_density <- ggplot(df.edit, aes(x = tOC, color = Diagnosis, fill = Diagnosis, name = "Diagnosis")) +
  geom_density(alpha = 0.3) +
  labs(title = "tOC density by diagnostic groups",
       x = "tOC",
       y = "Density") +
  theme_minimal() + 
  theme(
    text = element_text(family = "Arial", size = 14, colour = "black"),
    panel.grid = element_blank(),
    panel.background = element_blank(),
    axis.line = element_line(colour = "black"),
    axis.ticks = element_line(colour = "black"),
    axis.ticks.length = unit(0.2, "cm"),
    axis.text = element_text(colour = "black")) 

print(tOC_density)
```
# Perform pairwise comparisons on tOC using Mann Whitney Wilcoxon Rank Sum test to see signficances between groups
```{r}
pairwise_comparisons <- pairwise.wilcox.test(df$tOC, df$diagnosis, p.adjust.method = "fdr")
print(pairwise_comparisons)
```
#Combine Mean_thickness and tOC 
```{r}
Figure_2 <- plot_grid(
Mean_thickness, tOC_density,
  labels = "AUTO", nrow = 2)

print(Figure_2)
```

#----------------------------------------------           
## 3. Can the variance in tOC cortical thickness and subcortical volumes be explained by additional factors? 
#----------------------------------------------

# Question 1: Does sex have an impact on tOC, controlling for age? REJECTING THE LINEAR REGRESSION MODEL 
# Main source code on lmrob retrieved from: https://rdrr.io/rforge/robust/man/anova.lmRob.html
# All other robust linear regression models were adapted to 
```{r}
#Rob lin reg for sex by itself 
summary(lmrob(tOC ~ sex, data = df))
s1 <- lmrob(tOC ~ sex, data = df)
confint(s1) ##gives CI
rm(s1)

#Rob lin reg adjusted 
model_sex <- lmrob(tOC ~ sex + age, data = df)
summary(model_sex)
model_control <- lmrob(tOC ~ age, data = df)
anova(model_control, model_sex)
s2<-lmrob(tOC ~ sex + age, data = df)
confint(s2)
rm(s2)
#No sex differences observed for tOC, ANOVA and p-values for co-efficients were greater than 0.05.
```

# Question 2: Does diagnosis have an impact on tOC, controlling for age and sex?

#Robust linear regression 
#Interaction model

```{r}
#Rob lin reg for diagnosis by itself 
summary(lmrob(tOC ~ diagnosis, data = df))
s1 <- lmrob(tOC ~ diagnosis, data = df)
confint(s1) ##gives CI
rm(s1)

#Robust linear regression adjusted 
model_diagnosis <- lmrob(tOC ~ diagnosis + sex +  age, data = df)
summary(model_diagnosis)
model_control <- lmrob(tOC ~ age + sex, data = df)
anova_result <- anova(model_control, model_diagnosis)
print(anova_result)
s3<-lmrob(tOC ~ diagnosis + sex + age, data = df)
confint(s3)
rm(s3)
#Main result: adding diagnosis to the control model does not make much of a difference as the p-value comparing the models came out insignificant, greater than 0.05. So in terms of science, when adjusting for age and sex, diagnosis does not explain the variation or differences seen in tOC more than age and sex.

#FDR corrections (revise code)
robmodel_diagnosis <- lmrob(tOC ~ diagnosis + sex + age, data = df)
# Extract summary of the model
summary_robmodel_diagnosis <- summary(robmodel_diagnosis)

# FDR correction
p_values_table <- data.frame(
  Coefficients = rownames(summary_robmodel_diagnosis$coefficients),
  Original_P_Values = summary_robmodel_diagnosis$coefficients[, 4],
  FDR_Adjusted_P_Values = p.adjust(summary_robmodel_diagnosis$coefficients[, 4], method = "fdr")
)
# Display the table
print(p_values_table)
```

# Question 3: Does sex and diagnosis combined have an impact on tOC, controlling for age? Regression models for independent diagnosis
```{r}
# Subset the data for each diagnosis group
df_AD <- subset(df, diagnosis == "AD")
df_FTD <- subset(df, diagnosis == "Frontotemporal dementia")
df_MCI <- subset(df, diagnosis == "MCI")
df_Mixed <- subset(df, diagnosis == "Mixed dementia (AD and vascular dementia)")
df_Other <- subset(df, diagnosis == "Other non-AD Dementias")

# Run separate models
model_AD <- lmrob(tOC ~ sex + age, data = df_AD)
model_FTD <- lmrob(tOC ~ sex + age, data = df_FTD)
model_MCI <- lmrob(tOC ~ sex + age, data = df_MCI)
model_Mixed <- lmrob(tOC ~ sex + age, data = df_Mixed)
model_Other <- lmrob(tOC ~ sex + age, data = df_Other)

# Summarize models
summary(model_AD)
summary(model_FTD)
summary(model_MCI)
summary(model_Mixed)
summary(model_Other)
```

#Combined regression model
```{r}
#Rob lin reg for diagnosis and sex by itself 
summary(lmrob(tOC ~ diagnosis*sex, data = df))
s1 <- lmrob(tOC ~ diagnosis*sex, data = df)
confint(s1) ##gives CI
rm(s1)

#Robust linear regression for adjusted model 
model_sex_diagnosis <- lmrob(tOC ~ sex*diagnosis + age, data = df)
summary(model_sex_diagnosis)
model_control <- lmrob(tOC ~ age, data = df)
anova_result <- anova(model_control, model_sex_diagnosis)
print(anova_result)
s4<-lmrob(tOC ~ diagnosis*sex + age, data = df)
confint(s4)
rm(s4)

#FDR correction 
coeff_table <- summary(model_sex_diagnosis)$coefficients
p_values_table <- data.frame(
  Coefficients = rownames(coeff_table),
  Original_P_Values = coeff_table[, "Pr(>|t|)"],
  FDR_Adjusted_P_Values = p.adjust(coeff_table[, "Pr(>|t|)"], method = "fdr")
)

# Print the p-values table
print(p_values_table)
```
# Adjusted models
```{r}
#Rob lin reg for diagnosis and sex by itself 
summary(lmrob(tOC ~ diagnosis + sex, data = df))
s1 <- lmrob(tOC ~ diagnosis + sex, data = df)
confint(s1) ##gives CI
rm(s1)

#Robust linear regression for adjusted model 
model_sex_diagnosis <- lmrob(tOC ~ sex + diagnosis + age, data = df)
summary(model_sex_diagnosis)
model_control <- lmrob(tOC ~ age, data = df)
anova_result <- anova(model_control, model_sex_diagnosis)
print(anova_result)
s4<-lmrob(tOC ~ diagnosis + sex + age, data = df)
confint(s4)
rm(s4)

#FDR correction 
coeff_table <- summary(model_sex_diagnosis)$coefficients
p_values_table <- data.frame(
  Coefficients = rownames(coeff_table),
  Original_P_Values = coeff_table[, "Pr(>|t|)"],
  FDR_Adjusted_P_Values = p.adjust(coeff_table[, "Pr(>|t|)"], method = "fdr")
)

# Print the p-values table
print(p_values_table)
```
 
# Question 4: Do Years of education have an impact on tOC, controlling for age and sex? 
```{r}
#Rob lin reg for factor by itself 
summary(lmrob(tOC ~ Years.of.education, data = df))
s1 <- lmrob(tOC ~ Years.of.education, data = df)
confint(s1) ##gives CI
rm(s1)
```
#Adjusted model 
```{r}
#Robust linear regression for adjusted model 
model_education <- lmrob(tOC ~ Years.of.education + sex + age, data = df_edu_filtered)
summary(model_education)
model_control <- lmrob(tOC ~ age + sex, data = df_edu_filtered)
anova_result <- anova(model_control, model_education)
print(anova_result)
s5 <- lmrob(tOC ~ Years.of.education + sex + age, data = df_edu_filtered)
confint(s5)
rm(s5)

#FDR correction
coeff_table <- summary(model_education)$coefficients
p_values_education <- data.frame(
  Coefficients = rownames(coeff_table),
  Original_P_Values = coeff_table[, "Pr(>|t|)"],
  FDR_Adjusted_P_Values = p.adjust(coeff_table[, "Pr(>|t|)"], method = "fdr")
)

# Print the p-values table
print(p_values_education)
#Not significant results - education is not linked to the neuroanatomical variation. Why could this be? 
#So the FDR correction is not needed 
```

# Question 5: Does diagnosis and education have an impact on tOC, controlling for age and sex? 


```{r}
# Run separate models
# Fit the AD model using the MM method with increased iterations and adjusted tolerance
control_params <- lmrob.control(k.max = 500, refine.tol = 1e-06, max.it = 1000)
model_AD_MM <- lmrob(tOC ~ Years.of.education + sex + age, data = df_AD, method = "MM", control = control_params)
model_FTD <- lmrob(tOC ~ Years.of.education + sex + age, data = df_FTD)
model_MCI <- lmrob(tOC ~ Years.of.education + sex + age, data = df_MCI)
model_Mixed <- lmrob(tOC ~ Years.of.education + sex + age, data = df_Mixed)
model_Other <- lmrob(tOC ~ Years.of.education + sex + age, data = df_Other)

summary(model_AD_MM)
summary(model_FTD)
summary(model_MCI)
summary(model_Mixed)
summary(model_Other)
#Rob lin reg for factor by itself 
summary(lmrob(tOC ~ Years.of.education*diagnosis, data = df))
s1 <- lmrob(tOC ~ Years.of.education*diagnosis, data = df)
confint(s1) ##gives CI
rm(s1)

#Robust linear regression for adjusted model 
model_edu_diagnosis <- lmrob(tOC ~ Years.of.education*diagnosis + sex + age, data = df_edu_filtered)
summary(model_edu_diagnosis)
model_control <- lmrob(tOC ~ sex + age, data = df_edu_filtered)
anova_result <- anova(model_control, model_edu_diagnosis)
print(anova_result)
s6 <- lmrob(tOC ~ Years.of.education*diagnosis + sex + age, data = df_edu_filtered)
confint(s6)
rm(s6)
 
#FDR correction
coeff_table <- summary(model_edu_diagnosis)$coefficients
p_values_edu_diagnosis <- data.frame(
  Coefficients = rownames(coeff_table),
  Original_P_Values = coeff_table[, "Pr(>|t|)"],
  FDR_Adjusted_P_Values = p.adjust(coeff_table[, "Pr(>|t|)"], method = "fdr")
)

# Print the p-values table
print(p_values_edu_diagnosis)
```

#Adjusted model
```{r}
#Robust linear regression for adjusted model 
model_edu_diagnosis <- lmrob(tOC ~ Years.of.education + diagnosis + sex + age, data = df_edu_filtered)
summary(model_edu_diagnosis)
model_control <- lmrob(tOC ~ sex + age, data = df_edu_filtered)
anova_result <- anova(model_control, model_edu_diagnosis)
print(anova_result)
s6 <- lmrob(tOC ~ Years.of.education + diagnosis + sex + age, data = df_edu_filtered)
confint(s6)
rm(s6)
 
#FDR correction
coeff_table <- summary(model_edu_diagnosis)$coefficients
p_values_edu_diagnosis <- data.frame(
  Coefficients = rownames(coeff_table),
  Original_P_Values = coeff_table[, "Pr(>|t|)"],
  FDR_Adjusted_P_Values = p.adjust(coeff_table[, "Pr(>|t|)"], method = "fdr")
)

# Print the p-values table
print(p_values_edu_diagnosis)
```
#----------------------------------------------
# 4. 	Does this neuroanatomical variation (tOC) have an impact on cognitive scores? 
#----------------------------------------------
#Distribution of ACE_3 and ACE_R
```{r}
# Create the density plot
ggplot(df_ACE3_filtered, aes(x = ACE3_scores)) +
    geom_density(fill = "blue", alpha = 0.5) +
    labs(title = "Distribution of ACE3 Scores", x = "ACE3_scores") +
    theme_minimal()

# Create the density plot
ggplot(df_ACER_filtered, aes(x = ACER_scores)) +
    geom_density(fill = "red", alpha = 0.5) +
    labs(title = "Distribution of ACER Scores", x = "ACER_scores") +
    theme_minimal()
shapiro.test(df_ACE3_filtered$ACE3_scores)
shapiro.test(df_ACER_filtered$ACER_scores)
```
# ---------- TOTAL ACE3 ONLY ------------------

#1. Does tOC influence total ACE-3 scores in each diagnostic group??
```{r}
## Main effect
summary(lmrob(ACE3_scores ~ tOC, data = df_ACE3_filtered))
s1<- lmrob(ACE3_scores ~ tOC, data = df_ACE3_filtered)
confint(s1) ##gives CI

##interaction
full_model <- lmrob(ACE3_scores ~ tOC*diagnosis + age + sex, data = df_ACE3_filtered)
summary(full_model) ##outcome ~ predictor
control_model <- lmrob(ACE3_scores ~  age + sex, data = df_ACE3_filtered)
anova_result <- anova(control_model,full_model)
print(anova_result)
s2 <- lmrob(ACE3_scores ~ tOC*diagnosis + age + sex, data = df_ACE3_filtered)
confint(s2) 

#FDR correction
coeff_table <- summary(full_model)$coefficients
p_values_full_model <- data.frame(
  Coefficients = rownames(coeff_table),
  Original_P_Values = coeff_table[, "Pr(>|t|)"],
  FDR_Adjusted_P_Values = p.adjust(coeff_table[, "Pr(>|t|)"], method = "fdr")
)
# Print the p-values table
print(p_values_full_model)
```
#Adjusted effects
```{r}
## main effect adjusted 
Adjusted_total_ACE3 <- lmrob(ACE3_scores ~ tOC + diagnosis + age + sex, data = df_ACE3_filtered) ##outcome ~ predictor main effect

summary(Adjusted_total_ACE3) ##outcome ~ predictor
control_model <- lmrob(ACE3_scores ~  age + sex, data = df_ACE3_filtered)
anova_result <- anova(control_model,Adjusted_total_ACE3)
print(anova_result)
s2 <- lmrob(ACE3_scores ~ tOC + diagnosis + age + sex, data = df_ACE3_filtered)
confint(s2) 
rm(s2)

coeff_table <- summary(Adjusted_total_ACE3)$coefficients
p_values_adjusted <- data.frame(
  Coefficients = rownames(coeff_table),
  Original_P_Values = coeff_table[, "Pr(>|t|)"],
  FDR_Adjusted_P_Values = p.adjust(coeff_table[, "Pr(>|t|)"], method = "fdr")
)
# Print the p-values table
print(p_values_adjusted)
```

#2. 	Does Education and tOC have an impact on total ACE3_scores?
```{r}
## Main effect
summary(lmrob(ACE3_scores ~ tOC, data = df_ACE3_filtered))
s1<- lmrob(ACE3_scores ~ tOC, data = df_ACE3_filtered)
confint(s1) ##gives CI

##interaction
full_model <- lmrob(ACE3_scores ~ tOC*Years.of.education*diagnosis + age + sex, data = df_ACE3_filtered)
summary(full_model) ##outcome ~ predictor
control_model <- lmrob(ACE3_scores ~  age + sex, data = df_ACE3_filtered)
anova_result <- anova(control_model,full_model)
print(anova_result)
summary(full_model)
s3 <- lmrob(ACE3_scores ~ tOC*Years.of.education*diagnosis + age + sex, data = df_ACE3_filtered)
confint(s3) 
#FDR correction
coeff_table <- summary(full_model)$coefficients
p_values_full_model <- data.frame(
  Coefficients = rownames(coeff_table),
  Original_P_Values = coeff_table[, "Pr(>|t|)"],
  FDR_Adjusted_P_Values = p.adjust(coeff_table[, "Pr(>|t|)"], method = "fdr")
)
# Print the p-values table
print(p_values_full_model)
```

#Adjusted ACE3 Education
```{r}
## main effect adjusted 
Adjusted_ACE3_education <- lmrob(ACE3_scores ~ tOC + Years.of.education + diagnosis + age + sex, data = df_ACE3_filtered)##outcome ~ predictor main effect
summary(Adjusted_ACE3_education) ##outcome ~ predictor
control_model <- lmrob(ACE3_scores ~  age + sex, data = df_ACE3_filtered)
anova_result <- anova(control_model,Adjusted_ACE3_education)
print(anova_result)
s2 <- lmrob(ACE3_scores ~ tOC + Years.of.education + diagnosis + age + sex, data = df_ACE3_filtered)
confint(s2) 
rm(s2)

coeff_table <- summary(Adjusted_ACE3_education)$coefficients
p_values_adjusted_ACE3 <- data.frame(
  Coefficients = rownames(coeff_table),
  Original_P_Values = coeff_table[, "Pr(>|t|)"],
  FDR_Adjusted_P_Values = p.adjust(coeff_table[, "Pr(>|t|)"], method = "fdr")
)
# Print the p-values table
print(p_values_adjusted_ACE3)

```

# ---------- TOTAL ACER ONLY ------------------

#1. Does tOC influence total ACER scores in each diagnostic group??
```{r}
## Main effect
summary(lmrob(ACER_scores ~ tOC, data = df_ACER_filtered))
s1 <- lmrob(ACER_scores ~ tOC, data = df_ACER_filtered)
confint(s1) ##gives CI

## Interaction
full_model <- lmrob(ACER_scores ~ tOC * diagnosis + age + sex, data = df_ACER_filtered)
summary(full_model) ## outcome ~ predictor
control_model <- lmrob(ACER_scores ~ age + sex, data = df_ACER_filtered)
anova_result <- anova(control_model, full_model)
print(anova_result)
s5 <- lmrob(ACER_scores ~ tOC * diagnosis + age + sex, data = df_ACER_filtered)
confint(s5)

#FDR correction
coeff_table <- summary(full_model)$coefficients
p_values_full_model <- data.frame(
  Coefficients = rownames(coeff_table),
  Original_P_Values = coeff_table[, "Pr(>|t|)"],
  FDR_Adjusted_P_Values = p.adjust(coeff_table[, "Pr(>|t|)"], method = "fdr")
)
# Print the p-values table
print(p_values_full_model)
```

#Adjusted model
```{r}
## Main effect adjusted
full_model <- lmrob(ACER_scores ~ tOC + diagnosis + age + sex, data = df_ACER_filtered)
summary(lmrob(ACER_scores ~ tOC + diagnosis + age + sex, data = df_ACER_filtered)) ## outcome ~ predictor
control_model <- lmrob(ACER_scores ~ age + sex, data = df_ACER_filtered)
anova_result <- anova(control_model, full_model)
print(anova_result)
s5 <- lmrob(ACER_scores ~ tOC + diagnosis + age + sex, data = df_ACER_filtered)
confint(s5)

#FDR correction
coeff_table <- summary(full_model)$coefficients
p_values_full_model <- data.frame(
  Coefficients = rownames(coeff_table),
  Original_P_Values = coeff_table[, "Pr(>|t|)"],
  FDR_Adjusted_P_Values = p.adjust(coeff_table[, "Pr(>|t|)"], method = "fdr")
)
# Print the p-values table
print(p_values_full_model)
```

#2. 	Does Education and tOC have an impact on total ACER_scores?
```{r}
## Main effect
summary(lmrob(ACER_scores ~ tOC, data = df_ACER_filtered))
s1<- lmrob(ACER_scores ~ tOC, data = df_ACER_filtered)
confint(s1) ##gives CI

##interaction
full_model <- lmrob(ACER_scores ~ tOC*Years.of.education*diagnosis + age + sex, data = df_ACER_filtered)
summary(full_model) ##outcome ~ predictor
control_model <- lmrob(ACER_scores ~  age + sex, data = df_ACER_filtered)
anova_result <- anova(control_model,full_model)
print(anova_result)
summary(full_model)
s6 <- lmrob(ACER_scores ~ tOC*Years.of.education*diagnosis + age + sex, data = df_ACER_filtered)
confint(s6) 

#FDR correction
coeff_table <- summary(full_model)$coefficients
p_values_full_model <- data.frame(
  Coefficients = rownames(coeff_table),
  Original_P_Values = coeff_table[, "Pr(>|t|)"],
  FDR_Adjusted_P_Values = p.adjust(coeff_table[, "Pr(>|t|)"], method = "fdr")
)
# Print the p-values table
print(p_values_full_model)
```

#Adjusted model 
```{r}
## main effect adjusted 
full_model <- lmrob(ACER_scores ~ tOC + Years.of.education + diagnosis + age + sex, data = df_ACER_filtered) ##outcome ~ predictor main effect
summary(full_model)
control_model <- lmrob(ACER_scores ~  age + sex, data = df_ACER_filtered)
anova_result <- anova(control_model,full_model)
print(anova_result)
s6 <- lmrob(ACER_scores ~ tOC + Years.of.education + diagnosis + age + sex, data = df_ACER_filtered)
confint(s6) 

#FDR correction
coeff_table <- summary(full_model)$coefficients
p_values_full_model <- data.frame(
  Coefficients = rownames(coeff_table),
  Original_P_Values = coeff_table[, "Pr(>|t|)"],
  FDR_Adjusted_P_Values = p.adjust(coeff_table[, "Pr(>|t|)"], method = "fdr")
)
# Print the p-values table
print(p_values_full_model)
```

#Plot graphs for ACE-3 total scores 
```{r}
#Each group's diagnostic 
# Linear Model Plot
ggplot(df_ACE3_filtered, aes(x = tOC, y = ACE3_scores)) + 
  geom_point() +
  stat_smooth(method = "lm", col = "red", formula = y ~ x, size = 1)

# Quadratic Model Plot
ggplot(df_ACE3_filtered, aes(x = tOC, y = ACE3_scores, color = diagnosis)) + 
  geom_point() +
  stat_smooth(method = "lm", col = "red", formula = y ~ x + I(x^2), size = 1) 

# Separate Lines for Each Group
p <- df_ACE3_filtered %>% ggplot(aes(tOC, ACE3_scores)) 

p + 
  geom_point(aes(col = diagnosis)) + 
  geom_smooth(aes(col = diagnosis), method = 'lm', formula = y ~ x + I(x^2)) +
  theme(legend.background = element_rect(fill = "transparent"),
        legend.justification = c(0, 1),
        legend.position = c(0, 1))

# Separate Lines for Each Group (Enhanced Aesthetics)
p <- df_ACE3_filtered %>% ggplot(aes(tOC, ACE3_scores)) 
p + 
  geom_smooth(aes(col = diagnosis), method = 'lm', formula = y ~ x + I(x^2), fill = "snow2") +
  theme_minimal() +
  theme(text = element_text(family = "Times New Roman", face = "bold", size = 16)) +
  labs(colour = "Group", x = "Total Outlier Count (tOC)", y = "ACE3 Scores") +
  scale_color_manual(values = c("Non-AD Dementias" = "grey65", "MCI" = "mediumturquoise", "AD" = "lightslateblue"))

```
#Plot graphs for ACE-R total scores
```{r}
ggplot(df_ACER_filtered, aes(x = tOC, y = ACER_scores)) + 
  geom_point() +
  stat_smooth(method = "lm", col = "red", formula = y ~ x, size = 1)

# Quadratic Model Plot
ggplot(df_ACER_filtered, aes(x = tOC, y = ACER_scores, color = diagnosis)) + 
  geom_point() +
  stat_smooth(method = "lm", col = "red", formula = y ~ x + I(x^2), size = 1)

# Separate Lines for Each Group
p <- df_ACER_filtered %>% ggplot(aes(tOC, ACER_scores)) 

p + 
  geom_point(aes(col = diagnosis)) + 
  geom_smooth(aes(col = diagnosis), method = 'lm', formula = y ~ x + I(x^2)) +
  theme(legend.background = element_rect(fill = "transparent"),
        legend.justification = c(0, 1),
        legend.position = c(0, 1))

# Separate Lines for Each Group (Enhanced Aesthetics)
p <- df_ACER_filtered %>% ggplot(aes(tOC, ACER_scores)) 
p + 
  geom_smooth(aes(col = diagnosis), method = 'lm', formula = y ~ x + I(x^2), fill = "snow2") +
  theme_minimal() +
  theme(text = element_text(family = "Times New Roman", face = "bold", size = 16)) +
  labs(colour = "Group", x = "Total Outlier Count (tOC)", y = "ACER Scores") +
  scale_color_manual(values = c("Non-AD Dementias" = "grey65", "MCI" = "mediumturquoise", "AD" = "lightslateblue"))

```
# Combine all graphs into one 
```{r}
library(patchwork)

# Create individual plots using lmrob
p1 <- df_ACE3_filtered %>% 
  ggplot(aes(tOC, ACE3_scores)) + 
  geom_smooth(aes(col = diagnosis), method = 'lmrob', formula = y ~ x + I(x^2), fill = "snow2") +
  theme_minimal() +
  theme(
    text = element_text(family = "Arial", size = 16),
    panel.grid = element_blank(),
    panel.background = element_blank(),
    axis.line = element_line(colour = "black"),
    axis.ticks = element_line(colour = "black"),
    axis.ticks.length = unit(0.2, "cm")
  ) +
  labs(colour = "Group", x = "tOC", y = "Total ACE-III Scores") +
  scale_color_manual(values = c("Non-AD Dementias" = "grey65", "MCI" = "mediumturquoise", "AD" = "lightslateblue"))


p2 <- df_ACER_filtered %>% 
  ggplot(aes(tOC, ACER_scores)) + 
  geom_smooth(aes(col = diagnosis), method = 'lmrob', formula = y ~ x + I(x^2), fill = "snow2") +
  theme_minimal() +
  theme(
    text = element_text(family = "Arial", size = 16),
    panel.grid = element_blank(),
    panel.background = element_blank(),
    axis.line = element_line(colour = "black"),
    axis.ticks = element_line(colour = "black"),
    axis.ticks.length = unit(0.2, "cm")
  ) +
  labs(colour = "Group", x = "tOC", y = "Total ACE-R Scores") +
  scale_color_manual(values = c("Non-AD Dementias" = "grey65", "MCI" = "mediumturquoise", "AD" = "lightslateblue"))


# Combine plots into a grid with individual legends
final_plot1 <- plot_grid(p1, p2, nrow = 1, labels = "AUTO", label_size = 16)
print(final_plot1)
```
