==> 6. Individual Case Study Maps .Rmd <==
---
title: "Case_study_maps"
author: "Serena Verdi"
date: "26/10/2021"
output: html_document
---
Summary: This script creates individual brain maps according demographic, genotypic, cognitive and CSF amyloid critera

Paths to data will need to be changed accordingly.
#-----------------------------------------------
# Setup
#-----------------------------------------------
# Librarys
```{r setup}
library(dplyr)
library(psych)
library(tidyr)
library(readr)
library(data.table)
library(matrixStats)
library(stringr)
library(ggseg)
library(ggsegExtra)
library(ggsegDesterieux)
library(RPMG)
library(cowplot)
library(data.table)
library(e1071)
library(dplyr)
library(ggplot2)
library(plot.matrix)
library(proxy)
library(psych)
library(tidyr)
library(stringr)
library(RPMG)
library(viridis)
library(effectsize)
library(corrplot)
library(ggseg)
library(lme4)
library(lmtest)
library(arsenal)
library(lmerTest)
library(MuMIn)
library(effects)
library(ggeffects)
library(readr)
library(dplyr)
```

# Clear workspace
```{r}
rm(list = ls())
```
#Read in data 

```{r}
df <- read.csv("/Users/snehaprasanna/Desktop/Serena Verdi Project/QMIN/Final files/toc_df.csv")
df$diagnosis <- as.factor(df$diagnosis)
df$sex<- as.factor(df$sex)
df_ACE3_filtered <- read.csv("/Users/snehaprasanna/Desktop/Serena Verdi Project/QMIN/Final files/df_ACE3_filtered.csv")
df_ACER_filtered <- read.csv("/Users/snehaprasanna/Desktop/Serena Verdi Project/QMIN/Final files/df_ACER_filtered.csv")
df_edu_filtered <- read.csv("/Users/snehaprasanna/Desktop/Serena Verdi Project/QMIN/Final files/df_edu_filtered.csv")

```
# Prepare data
```{r}
#df <- df[-grep("subject_id|.x|X|VISCODE|USER|SCAN|diff|EXAM|date|RUNDATE|VERSION|ORIG|STATUS|LONI|IMAGE|quality",names(df))]

df$sex <- gsub("2", "Female", df$sex)
df$sex <- gsub("1", "Male", df$sex)
df$sex <- factor(df$sex)
```


#-----------------------------------------------
#Create total outlier score
#-----------------------------------------------
# Thresholding z-scores 

```{r}
#Filter Cortical Thicknesses
df_ct <- df[, -c(183:202)] # need to check as manually selecting columns here. Alt way could grep 
# find out length to work out % of outliers in group
n <- length(unique(df_ct$RID))

#Filter Subcortical Volumes 
df_sv <- df[, -c(203:350)] # need to check as manually selecting columns here. Alt way could grep 
# find out length to work out % of outliers in group
n <- length(unique(df_sv$RID))
```
#-----------------------------------------------
# Filter critera
#-----------------------------------------------
# Group, Age, sex, Years of education, 
```{r}
# Filter data based on the specified criteria
subset_AD <- df %>%
  filter(diagnosis == "AD") %>%
  filter(sex == "Female") %>%
  filter(between(Years.of.education, 10, 15)) %>%
  filter(between(age, 74.0, 75.0)) %>%
  filter(ACE3_scores < 82)
subset_AD <- subset_AD %>% 
  filter(!(RID %in% c(33128, 33478)))

# Display the RID column of the resulting subset
subset_AD$RID

subset_FTD <- df %>%
  filter(diagnosis == "Frontotemporal dementia") %>%
  filter(sex == "Male") %>%
  filter(between(Years.of.education, 10, 15)) %>%
  filter(between(age, 72.0, 75.5)) %>%
  filter(ACER_scores <= 82 | ACE3_scores <= 82)
# Display the RID column of the resulting subset
subset_FTD$RID

subset_Mixed <- df %>%
  filter(diagnosis == "Mixed dementia (AD and vascular dementia)") %>%
  filter(sex == "Male") %>%
  filter(between(Years.of.education, 10, 15)) %>%
  filter(between(age, 80.0, 84.1)) %>%
  filter(ACER_scores <= 82 | ACE3_scores <= 82)
# Display the RID column of the resulting subset
subset_Mixed$RID


subset_Vasc <- df %>%
  filter(diagnosis == "Other non-AD Dementias") %>%
  filter(sex == "Male") %>%
  filter(between(Years.of.education, 10, 15)) %>%
  filter(between(age, 75.0, 79.0)) %>%
  filter(ACER_scores <= 82 | ACE3_scores <= 82)
# Display the RID column of the resulting subset
subset_Vasc$RID

# Remove rows with RID 31700 and 33941
subset_Vasc <- subset_Vasc %>% 
  filter(!(RID %in% c(31700, 33941)))
```

#-----------------------------------------------
# AD Individual Cortical Thickness Maps 
#-----------------------------------------------

#Subject 1 AD
```{r}
subject1_AD <- df_ct %>% filter(RID == "32651")
subject1_AD <- subject1_AD[,grep("_bin", names(subject1_AD))]
subject1_AD <- gather(subject1_AD)
subject1_AD$ROI <- subject1_AD$key
subject1_AD$ROI <- gsub("_bin", "", subject1_AD$ROI)
subject1_AD <- subject1_AD

data(desterieux)

left.subject1_AD <- subject1_AD %>%
  filter(str_detect(ROI, "lh"))
x <- gsub("lh_", "", left.subject1_AD$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
list_matches <- as.data.frame(compare_lists[,1] %in% compare_lists[,2])
#
compare_lists[!list_matches,]
## if no mismatches, than add to data.frame as region

left.subject1_AD$region <- renamed_ROIs

right.subject1_AD <- subject1_AD %>%
  filter(str_detect(ROI, "rh"))
x <- gsub("rh_", "", right.subject1_AD$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))

desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "right"))$region
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
compare_lists[!list_matches,]

## if no mismatches, than add to data.frame as region
right.subject1_AD$region <- renamed_ROIs

# create df atlas
desterieux <- as.data.frame(desterieux)

# divide desterieux into left and right
desterieux_r <- desterieux %>% filter(hemi == "right")
desterieux_l <- desterieux %>% filter(hemi == "left")

desterieux_r <- desterieux_r[grep("atlas|region|side|roi|label",names(desterieux_r))]

desterieux_l <- desterieux_l[grep("atlas|region|side|roi|label",names(desterieux_l))]

# merge left and right hemi outlier dfs respectively with the desterieux left and righ dfs
left.subject1_AD <- merge(desterieux_l, left.subject1_AD, by= "region", all.x = TRUE) 

right.subject1_AD <- merge(desterieux_r, right.subject1_AD, by= "region", all.x = TRUE) 

# bind together 
subject1_AD <- rbind(left.subject1_AD,right.subject1_AD)

rm(desterieux)

# Modify the plot commands to hide the legends
left <- ggseg(.data=left.subject1_AD, atlas = desterieux, mapping=aes(fill=value), hemisphere = "left", colour = "white", size = 0.2) + 
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

right <- ggseg(.data=right.subject1_AD, atlas = desterieux, mapping=aes(fill=value), hemisphere = "right", colour = "white", size = 0.2) +
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Combine the plots without legends
brain_plots <- plot_grid(left, right, nrow = 1)

# Add title close to the left of the first image of the left hemisphere
subject1_AD_ct <- ggdraw() +
  draw_plot(brain_plots) +
  draw_label("Patient 1", fontface = 'bold', x = 0.1, y = 0.95, hjust = 0, size = 14)

print(subject1_AD_ct)
```

#Subject 2 AD

```{r}
subject2_AD <- df_ct %>% filter(RID == "36460")
subject2_AD <- subject2_AD[,grep("_bin", names(subject2_AD))]
subject2_AD <- gather(subject2_AD)
subject2_AD$ROI <- subject2_AD$key
subject2_AD$ROI <- gsub("_bin", "", subject2_AD$ROI)
subject2_AD <- subject2_AD

data(desterieux)

left.subject2_AD <- subject2_AD %>%
  filter(str_detect(ROI, "lh"))
x <- gsub("lh_", "", left.subject2_AD$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
list_matches <- as.data.frame(compare_lists[,1] %in% compare_lists[,2])
#
compare_lists[!list_matches,]
## if no mismatches, than add to data.frame as region

left.subject2_AD$region <- renamed_ROIs

right.subject2_AD <- subject2_AD %>%
  filter(str_detect(ROI, "rh"))
x <- gsub("rh_", "", right.subject2_AD$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))

desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "right"))$region
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
compare_lists[!list_matches,]

## if no mismatches, than add to data.frame as region
right.subject2_AD$region <- renamed_ROIs

# create df atlas
desterieux <- as.data.frame(desterieux)

# divide desterieux into left and right
desterieux_r <- desterieux %>% filter(hemi == "right")
desterieux_l <- desterieux %>% filter(hemi == "left")

desterieux_r <- desterieux_r[grep("atlas|region|side|roi|label",names(desterieux_r))]

desterieux_l <- desterieux_l[grep("atlas|region|side|roi|label",names(desterieux_l))]

# merge left and right hemi outlier dfs respectively with the desterieux left and righ dfs
left.subject2_AD <- merge(desterieux_l, left.subject2_AD, by= "region", all.x = TRUE) 

right.subject2_AD <- merge(desterieux_r, right.subject2_AD, by= "region", all.x = TRUE) 

# bind together 
subject2_AD <- rbind(left.subject2_AD,right.subject2_AD)

rm(desterieux)

# Modify the plot commands to hide the legends
left <- ggseg(.data=left.subject2_AD, atlas = desterieux, mapping=aes(fill=value), hemisphere = "left", colour = "white", size = 0.2) + 
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

right <- ggseg(.data=right.subject2_AD, atlas = desterieux, mapping=aes(fill=value), hemisphere = "right", colour = "white", size = 0.2) +
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Combine the plots without legends
brain_plots <- plot_grid(left, right, nrow = 1)

# Add title close to the left of the first image of the left hemisphere
subject2_AD_ct <- ggdraw() +
  draw_plot(brain_plots) +
  draw_label("Patient 2", fontface = 'bold', x = 0.1, y = 0.95, hjust = 0, size = 14)

print(subject2_AD_ct)
```

# Combine cortical thickness plots 

```{r}
# Combine the two plots
combined_AD_ct <- plot_grid(subject1_AD_ct, subject2_AD_ct, nrow = 2, labels = NULL)

# Print the final combined plot
print(combined_AD_ct)

# Sample data for creating the legend (adjust as necessary)
legend_text <- data.frame(
  x = 1,
  y = 1,
  text = "Diagnosis: Alzheimer's Disease\nSex: Female\nAge: 74–75\nEducation Years: 11-13\nTotal ACE-III: 55–56"
)

# Create the legend plot with a grey background
legend_plot_AD <- ggplot(legend_text, aes(x, y, label = text)) +
  geom_label(size = 5, hjust = 0, fill = "grey90", label.padding = unit(0.5, "lines")) +
  theme_void() +
  theme(plot.margin = margin(0, 0, 10, 0))

# Adjust the width of the plot grid to bring the legend boxes closer
combined_AD_ct <- plot_grid(
  legend_plot_AD,
  combined_AD_ct, 
  ncol = 2,  # Arrange the plots in a single row
  rel_widths = c(2.5, 4)  # Adjust the relative widths if necessary
)

# Add the label "A" to the plot
combined_AD_ct <- ggdraw() +
  draw_plot(combined_AD_ct) +
 draw_label("A", x = 0.15, y = 0.95, size = 18, fontface = "bold")

# Print the final combined plot with the legend and label
print(combined_AD_ct)

```

#-----------------------------------------------
# FTD Individual Cortical Thickness Maps 
#-----------------------------------------------

#Subject 1 FTD
```{r}
subject1_FTD_ct <- df_ct %>% filter(RID == "31108")
subject1_FTD_ct <- subject1_FTD_ct[,grep("_bin", names(subject1_FTD_ct))]
subject1_FTD_ct <- gather(subject1_FTD_ct)
subject1_FTD_ct$ROI <- subject1_FTD_ct$key
subject1_FTD_ct$ROI <- gsub("_bin", "", subject1_FTD_ct$ROI)
subject1_FTD_ct

data(desterieux)

left.subject1_FTD <- subject1_FTD_ct %>%
  filter(str_detect(ROI, "lh"))
x <- gsub("lh_", "", left.subject1_FTD$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
list_matches <- as.data.frame(compare_lists[,1] %in% compare_lists[,2])
#
compare_lists[!list_matches,]
## if no mismatches, than add to data.frame as region

left.subject1_FTD$region <- renamed_ROIs

right.subject1_FTD <- subject1_FTD_ct %>%
  filter(str_detect(ROI, "rh"))
x <- gsub("rh_", "", right.subject1_FTD$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))

desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "right"))$region
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
compare_lists[!list_matches,]

## if no mismatches, than add to data.frame as region
right.subject1_FTD$region <- renamed_ROIs

# create df atlas
desterieux <- as.data.frame(desterieux)

# divide desterieux into left and right
desterieux_r <- desterieux %>% filter(hemi == "right")
desterieux_l <- desterieux %>% filter(hemi == "left")

desterieux_r <- desterieux_r[grep("atlas|region|side|roi|label",names(desterieux_r))]

desterieux_l <- desterieux_l[grep("atlas|region|side|roi|label",names(desterieux_l))]

# merge left and right hemi outlier dfs respectively with the desterieux left and righ dfs
left.subject1_FTD <- merge(desterieux_l, left.subject1_AD, by= "region", all.x = TRUE) 

right.subject1_FTD <- merge(desterieux_r, right.subject1_AD, by= "region", all.x = TRUE) 

# bind together 
subject1_FTD <- rbind(left.subject1_FTD,right.subject1_FTD)

rm(desterieux)

# Modify the plot commands to hide the legends
left <- ggseg(.data=left.subject1_FTD, atlas = desterieux, mapping=aes(fill=value), hemisphere = "left", colour = "white", size = 0.2) + 
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

right <- ggseg(.data=right.subject1_FTD, atlas = desterieux, mapping=aes(fill=value), hemisphere = "right", colour = "white", size = 0.2) +
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Combine the plots without legends
brain_plots <- plot_grid(left, right, nrow = 1)

# Add title close to the left of the first image of the left hemisphere
subject1_FTD_ct <- ggdraw() +
  draw_plot(brain_plots) +
  draw_label("Patient 1", fontface = 'bold', x = 0.1, y = 0.9, hjust = 0, size = 14)

print(subject1_FTD_ct)
```

#Subject 2 FTD

```{r}
subject2_FTD_ct <- df_ct %>% filter(RID == "33401")
subject2_FTD_ct <- subject2_FTD_ct[,grep("_bin", names(subject2_FTD_ct))]
subject2_FTD_ct <- gather(subject2_FTD_ct)
subject2_FTD_ct$ROI <- subject2_FTD_ct$key
subject2_FTD_ct$ROI <- gsub("_bin", "", subject2_FTD_ct$ROI)
subject2_FTD_ct

data(desterieux)

left.subject2_FTD <- subject2_FTD_ct %>%
  filter(str_detect(ROI, "lh"))
x <- gsub("lh_", "", left.subject2_FTD$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
list_matches <- as.data.frame(compare_lists[,1] %in% compare_lists[,2])
#
compare_lists[!list_matches,]
## if no mismatches, then add to data.frame as region

left.subject2_FTD$region <- renamed_ROIs

right.subject2_FTD <- subject2_FTD_ct %>%
  filter(str_detect(ROI, "rh"))
x <- gsub("rh_", "", right.subject2_FTD$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))

desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "right"))$region
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
compare_lists[!list_matches,]

## if no mismatches, then add to data.frame as region
right.subject2_FTD$region <- renamed_ROIs

# create df atlas
desterieux <- as.data.frame(desterieux)

# divide desterieux into left and right
desterieux_r <- desterieux %>% filter(hemi == "right")
desterieux_l <- desterieux %>% filter(hemi == "left")

desterieux_r <- desterieux_r[grep("atlas|region|side|roi|label",names(desterieux_r))]

desterieux_l <- desterieux_l[grep("atlas|region|side|roi|label",names(desterieux_l))]

# merge left and right hemi outlier dfs respectively with the desterieux left and right dfs
left.subject2_FTD <- merge(desterieux_l, left.subject2_FTD, by= "region", all.x = TRUE) 

right.subject2_FTD <- merge(desterieux_r, right.subject2_FTD, by= "region", all.x = TRUE) 

# bind together 
subject2_FTD <- rbind(left.subject2_FTD, right.subject2_FTD)

rm(desterieux)

# Modify the plot commands to hide the legends
left <- ggseg(.data=left.subject2_FTD, atlas = desterieux, mapping=aes(fill=value), hemisphere = "left", colour = "white", size = 0.2) + 
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

right <- ggseg(.data=right.subject2_FTD, atlas = desterieux, mapping=aes(fill=value), hemisphere = "right", colour = "white", size = 0.2) +
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Combine the plots without legends
brain_plots <- plot_grid(left, right, nrow = 1)

# Add title close to the left of the first image of the left hemisphere
subject2_FTD_ct <- ggdraw() +
  draw_plot(brain_plots) +
  draw_label("Patient 2", fontface = 'bold', x = 0.1, y = 0.9, hjust = 0, size = 14)

print(subject2_FTD_ct)
```

# Combine cortical thickness plots 

```{r}

# Combine the two plots
combined_FTD_ct <- plot_grid(subject1_FTD_ct, subject2_FTD_ct, nrow = 2, labels = NULL)

# Print the final combined plot
print(combined_FTD_ct)

# Sample data for creating the legend (adjust as necessary)
legend_text <- data.frame(
  x = 1,
  y = 1,
  text = "Diagnosis: Frontotemporal Dementia\nSex: Male\nAge: 72–75\nEducation Years: 10-13\nTotal ACE-R: 70–76"
)

# Create the legend plot with a grey background
legend_plot_FTD <- ggplot(legend_text, aes(x, y, label = text)) +
  geom_label(size = 5, hjust = 0, fill = "grey90", label.padding = unit(0.5, "lines")) +
  theme_void() +
  theme(plot.margin = margin(0, 0, 10, 0))

# Adjust the width of the plot grid to bring the legend boxes closer
combined_FTD_ct <- plot_grid(
  legend_plot_FTD,
  combined_FTD_ct, 
  ncol = 2,  # Arrange the plots in a single row
  rel_widths = c(3.2, 4)  # Adjust the relative widths if necessary
)

# Add the label "B" to the plot
combined_FTD_ct <- ggdraw() +
  draw_plot(combined_FTD_ct) +
  draw_label("B", x = 0.15, y = 0.95, size = 18, fontface = "bold")

# Print the final combined plot with the legend and label
print(combined_FTD_ct) 
```
#-----------------------------------------------
# Mixed dementia Individual Cortical Thickness Maps 
#-----------------------------------------------

#Patient 1 Mixed dementia
```{r}
subject1_Mixed_ct <- df_ct %>% filter(RID == "32934")
subject1_Mixed_ct <- subject1_Mixed_ct[,grep("_bin", names(subject1_Mixed_ct))]
subject1_Mixed_ct <- gather(subject1_Mixed_ct)
subject1_Mixed_ct$ROI <- subject1_Mixed_ct$key
subject1_Mixed_ct$ROI <- gsub("_bin", "", subject1_Mixed_ct$ROI)
subject1_Mixed_ct

data(desterieux)

left.subject1_Mixed <- subject1_Mixed_ct %>%
  filter(str_detect(ROI, "lh"))
x <- gsub("lh_", "", left.subject1_Mixed$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
list_matches <- as.data.frame(compare_lists[,1] %in% compare_lists[,2])
#
compare_lists[!list_matches,]
## if no mismatches, than add to data.frame as region

left.subject1_Mixed$region <- renamed_ROIs

right.subject1_Mixed <- subject1_Mixed_ct %>%
  filter(str_detect(ROI, "rh"))
x <- gsub("rh_", "", right.subject1_Mixed$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))

desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "right"))$region
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
compare_lists[!list_matches,]

## if no mismatches, than add to data.frame as region
right.subject1_Mixed$region <- renamed_ROIs

# create df atlas
desterieux <- as.data.frame(desterieux)

# divide desterieux into left and right
desterieux_r <- desterieux %>% filter(hemi == "right")
desterieux_l <- desterieux %>% filter(hemi == "left")

desterieux_r <- desterieux_r[grep("atlas|region|side|roi|label",names(desterieux_r))]

desterieux_l <- desterieux_l[grep("atlas|region|side|roi|label",names(desterieux_l))]

# merge left and right hemi outlier dfs respectively with the desterieux left and righ dfs
left.subject1_Mixed <- merge(desterieux_l, left.subject1_Mixed, by= "region", all.x = TRUE) 

right.subject1_Mixed <- merge(desterieux_r, right.subject1_Mixed, by= "region", all.x = TRUE) 

# bind together 
subject1_Mixed <- rbind(left.subject1_Mixed,right.subject1_Mixed)

rm(desterieux)

# Modify the plot commands to hide the legends
left <- ggseg(.data=left.subject1_Mixed, atlas = desterieux, mapping=aes(fill=value), hemisphere = "left", colour = "white", size = 0.2) + 
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

right <- ggseg(.data=right.subject1_Mixed, atlas = desterieux, mapping=aes(fill=value), hemisphere = "right", colour = "white", size = 0.2) +
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Combine the plots without legends
brain_plots <- plot_grid(left, right, nrow = 1)

# Add title close to the left of the first image of the left hemisphere
subject1_Mixed_ct <- ggdraw() +
  draw_plot(brain_plots) +
  draw_label("Patient 1", fontface = 'bold', x = 0.1, y = 0.9, hjust = 0, size = 14)

print(subject1_Mixed_ct)
```

#Patient 2 Mixed dementia

```{r}
subject2_Mixed_ct <- df_ct %>% filter(RID == "35015")
subject2_Mixed_ct <- subject2_Mixed_ct[,grep("_bin", names(subject2_Mixed_ct))]
subject2_Mixed_ct <- gather(subject2_Mixed_ct)
subject2_Mixed_ct$ROI <- subject2_Mixed_ct$key
subject2_Mixed_ct$ROI <- gsub("_bin", "", subject2_Mixed_ct$ROI)
subject2_Mixed_ct

data(desterieux)

left.subject2_Mixed <- subject2_Mixed_ct %>%
  filter(str_detect(ROI, "lh"))
x <- gsub("lh_", "", left.subject2_Mixed$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
list_matches <- as.data.frame(compare_lists[,1] %in% compare_lists[,2])
#
compare_lists[!list_matches,]
## if no mismatches, than add to data.frame as region

left.subject2_Mixed$region <- renamed_ROIs

right.subject2_Mixed <- subject2_Mixed_ct %>%
  filter(str_detect(ROI, "rh"))
x <- gsub("rh_", "", right.subject2_Mixed$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))

desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "right"))$region
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
compare_lists[!list_matches,]

## if no mismatches, than add to data.frame as region
right.subject2_Mixed$region <- renamed_ROIs

# create df atlas
desterieux <- as.data.frame(desterieux)

# divide desterieux into left and right
desterieux_r <- desterieux %>% filter(hemi == "right")
desterieux_l <- desterieux %>% filter(hemi == "left")

desterieux_r <- desterieux_r[grep("atlas|region|side|roi|label",names(desterieux_r))]

desterieux_l <- desterieux_l[grep("atlas|region|side|roi|label",names(desterieux_l))]

# merge left and right hemi outlier dfs respectively with the desterieux left and right dfs
left.subject2_Mixed <- merge(desterieux_l, left.subject2_Mixed, by= "region", all.x = TRUE) 

right.subject2_Mixed <- merge(desterieux_r, right.subject2_Mixed, by= "region", all.x = TRUE) 

# bind together 
subject2_Mixed <- rbind(left.subject2_Mixed,right.subject2_Mixed)

rm(desterieux)

# Modify the plot commands to hide the legends
left <- ggseg(.data=left.subject2_Mixed, atlas = desterieux, mapping=aes(fill=value), hemisphere = "left", colour = "white", size = 0.2) + 
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

right <- ggseg(.data=right.subject2_Mixed, atlas = desterieux, mapping=aes(fill=value), hemisphere = "right", colour = "white", size = 0.2) +
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Combine the plots without legends
brain_plots <- plot_grid(left, right, nrow = 1)

# Add title close to the left of the first image of the left hemisphere
subject2_Mixed_ct <- ggdraw() +
  draw_plot(brain_plots) +
  draw_label("Patient 2", fontface = 'bold', x = 0.1, y = 0.9, hjust = 0, size = 14)

print(subject2_Mixed_ct)
```

#Combined Mixed ct plots

```{r}
# Combine the two plots for Mixed dementia
combined_Mixed_ct <- plot_grid(subject1_Mixed_ct, subject2_Mixed_ct, nrow = 2, labels = NULL)

# Print the final combined plot
print(combined_Mixed_ct)

# Sample data for creating the legend (adjust as necessary)
legend_text <- data.frame(
  x = 1,
  y = 1,
  text = "Diagnosis: Mixed Dementia\nSex: Male\nAge: 80–84\nEducation Years: 11\nTotal ACE-III: 62–78"
)

# Create the legend plot with a grey background
legend_plot_Mixed <- ggplot(legend_text, aes(x, y, label = text)) +
  geom_label(size = 5, hjust = 0, fill = "grey90", label.padding = unit(0.5, "lines")) +
  theme_void() +
  theme(plot.margin = margin(0, 0, 10, 0))

# Adjust the width of the plot grid to bring the legend boxes closer
combined_Mixed_ct <- plot_grid(
  legend_plot_Mixed,
  combined_Mixed_ct, 
  ncol = 2,  # Arrange the plots in a single row
  rel_widths = c(2.5, 4)  # Adjust the relative widths if necessary
)

# Add the label "A" to the plot
combined_Mixed_ct <- ggdraw() +
  draw_plot(combined_Mixed_ct) +
  draw_label("C", x = 0.15, y = 0.9, size = 18, fontface = "bold")

# Print the final combined plot with the legend and label
print(combined_Mixed_ct)

```

#-----------------------------------------------
# Vascular dementia Individual Cortical Thickness Maps 
#-----------------------------------------------

#Subject 1 Vascular dementia
```{r}
subject1_Vasc_ct <- df_ct %>% filter(RID == "33770")
subject1_Vasc_ct <- subject1_Vasc_ct[,grep("_bin", names(subject1_Vasc_ct))]
subject1_Vasc_ct <- gather(subject1_Vasc_ct)
subject1_Vasc_ct$ROI <- subject1_Vasc_ct$key
subject1_Vasc_ct$ROI <- gsub("_bin", "", subject1_Vasc_ct$ROI)
subject1_Vasc_ct

data(desterieux)

left.subject1_Vasc <- subject1_Vasc_ct %>%
  filter(str_detect(ROI, "lh"))
x <- gsub("lh_", "", left.subject1_Vasc$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
list_matches <- as.data.frame(compare_lists[,1] %in% compare_lists[,2])
#
compare_lists[!list_matches,]
## if no mismatches, than add to data.frame as region

left.subject1_Vasc$region <- renamed_ROIs

right.subject1_Vasc <- subject1_Vasc_ct %>%
  filter(str_detect(ROI, "rh"))
x <- gsub("rh_", "", right.subject1_Vasc$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))

desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "right"))$region
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
compare_lists[!list_matches,]

## if no mismatches, than add to data.frame as region
right.subject1_Vasc$region <- renamed_ROIs

# create df atlas
desterieux <- as.data.frame(desterieux)

# divide desterieux into left and right
desterieux_r <- desterieux %>% filter(hemi == "right")
desterieux_l <- desterieux %>% filter(hemi == "left")

desterieux_r <- desterieux_r[grep("atlas|region|side|roi|label",names(desterieux_r))]

desterieux_l <- desterieux_l[grep("atlas|region|side|roi|label",names(desterieux_l))]

# merge left and right hemi outlier dfs respectively with the desterieux left and righ dfs
left.subject1_Vasc <- merge(desterieux_l, left.subject1_Vasc, by= "region", all.x = TRUE) 

right.subject1_Vasc <- merge(desterieux_r, right.subject1_Vasc, by= "region", all.x = TRUE) 

# bind together 
subject1_Vasc <- rbind(left.subject1_Vasc,right.subject1_Vasc)

rm(desterieux)

# Modify the plot commands to hide the legends
left <- ggseg(.data=left.subject1_Vasc, atlas = desterieux, mapping=aes(fill=value), hemisphere = "left", colour = "white", size = 0.2) + 
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

right <- ggseg(.data=right.subject1_Vasc, atlas = desterieux, mapping=aes(fill=value), hemisphere = "right", colour = "white", size = 0.2) +
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Combine the plots without legends
brain_plots <- plot_grid(left, right, nrow = 1)

# Add title close to the left of the first image of the left hemisphere
subject1_Vasc_ct <- ggdraw() +
  draw_plot(brain_plots) +
  draw_label("Patient 1", fontface = 'bold', x = 0.1, y = 0.9, hjust = 0, size = 14)

print(subject1_Vasc_ct)
```

#Subject 2 Vascular dementia

```{r}
subject2_Vasc_ct <- df_ct %>% filter(RID == "34493")
subject2_Vasc_ct <- subject2_Vasc_ct[,grep("_bin", names(subject2_Vasc_ct))]
subject2_Vasc_ct <- gather(subject2_Vasc_ct)
subject2_Vasc_ct$ROI <- subject2_Vasc_ct$key
subject2_Vasc_ct$ROI <- gsub("_bin", "", subject2_Vasc_ct$ROI)
subject2_Vasc_ct

data(desterieux)

left.subject2_Vasc <- subject2_Vasc_ct %>%
  filter(str_detect(ROI, "lh"))
x <- gsub("lh_", "", left.subject2_Vasc$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
list_matches <- as.data.frame(compare_lists[,1] %in% compare_lists[,2])
#
compare_lists[!list_matches,]
## if no mismatches, then add to data.frame as region

left.subject2_Vasc$region <- renamed_ROIs

right.subject2_Vasc <- subject2_Vasc_ct %>%
  filter(str_detect(ROI, "rh"))
x <- gsub("rh_", "", right.subject2_Vasc$ROI)
y <- gsub("G.S_", "G and S ", x)
z <- gsub("G_", "G ", y)
z1 <- gsub("S_", "S ", z)
z2 <- gsub("_", " ", z1)
z3 <- gsub(" bin", "", z2)
z4 <- gsub("\\.", " ", z3)
z5 <- gsub("cingul ", "cingul-", z4)
z6 <- gsub("Mid ", "Mid-", z5)
z7 <- gsub("Post ", "Post", z6)
z8 <- gsub("inf ", "inf-", z7)
z9 <- gsub("med ", "med-", z8)
z10 <- gsub("sup ", "sup-", z9)
z11 <- gsub("Fis ant ", "Fis-ant-", z10)
z12 <- gsub("precentral ", "precentral-", z11)
z13 <- gsub("Fis pos ", "Fis-pos-", z12)
z14 <- gsub("lg S", "lg and S", z13)
z15 <- gsub("oc temp", "oc-temp", z14)
z16 <- gsub("sup and transversal", "sup-transversal", z15)
z17 <- gsub("orbital H Shaped", "orbital-H Shaped", z16)
z18 <- gsub("oc sup-transversal", "oc sup and transversal", z17)
z19 <- gsub("prim Jensen", "prim-Jensen", z18)
z20 <- gsub("S oc-temp med-Lingual", "S oc-temp med and Lingual", z19)
z21 <- gsub("lat fusifor", "lat-fusifor", z20)
z22 <- gsub("middle Lunatus", "middle and Lunatus", z21)
z23 <- gsub("intrapariet P trans", "intrapariet and P trans", z22)
z24 <- gsub("thickness", "", z23)
z25 <- gsub("sup- ", "sup", z24)
z26 <- gsub("inf- ", "inf", z25)
z27 <- gsub("Postdorsal", "Post-dorsal", z26)
z28 <- gsub("Postventral", "Post-ventral", z27)
z29 <- gsub("G precentral-", "G precentral", z28)
renamed_ROIs <- gsub("Lat Fis post", "Lat Fis-post", z29)
renamed_ROIs <- as.data.frame(renamed_ROIs)

renamed_ROIs <- str_trim(renamed_ROIs$renamed_ROIs, side = c("right"))

desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "right"))$region
desterieux <- as.data.frame(desterieux)
desterieux_ROIs <- as.data.frame(desterieux %>% filter(hemi == "left"))$region
compare_lists <- cbind(sort(renamed_ROIs), sort(unique(desterieux_ROIs)))
list_matches <- compare_lists[,1] %in% compare_lists[,2]
compare_lists[!list_matches,]

## if no mismatches, then add to data.frame as region
right.subject2_Vasc$region <- renamed_ROIs

# create df atlas
desterieux <- as.data.frame(desterieux)

# divide desterieux into left and right
desterieux_r <- desterieux %>% filter(hemi == "right")
desterieux_l <- desterieux %>% filter(hemi == "left")

desterieux_r <- desterieux_r[grep("atlas|region|side|roi|label",names(desterieux_r))]

desterieux_l <- desterieux_l[grep("atlas|region|side|roi|label",names(desterieux_l))]

# merge left and right hemi outlier dfs respectively with the desterieux left and right dfs
left.subject2_Vasc <- merge(desterieux_l, left.subject2_Vasc, by= "region", all.x = TRUE) 

right.subject2_Vasc <- merge(desterieux_r, right.subject2_Vasc, by= "region", all.x = TRUE) 

# bind together 
subject2_Vasc <- rbind(left.subject2_Vasc,right.subject2_Vasc)

rm(desterieux)

# Modify the plot commands to hide the legends
left <- ggseg(.data=left.subject2_Vasc, atlas = desterieux, mapping=aes(fill=value), hemisphere = "left", colour = "white", size = 0.2) + 
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

right <- ggseg(.data=right.subject2_Vasc, atlas = desterieux, mapping=aes(fill=value), hemisphere = "right", colour = "white", size = 0.2) +
  scale_fill_gradientn(limits = c(0.1, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Combine the plots without legends
brain_plots <- plot_grid(left, right, nrow = 1)

# Add title close to the left of the first image of the left hemisphere
subject2_Vasc_ct <- ggdraw() +
  draw_plot(brain_plots) +
  draw_label("Patient 2", fontface = 'bold', x = 0.1, y = 0.9, hjust = 0, size = 14)

print(subject2_Vasc_ct)

```

# Combine cortical thickness plots 

```{r}

# Combine the two plots for Vascular dementia
combined_Vasc_ct <- plot_grid(subject1_Vasc_ct, subject2_Vasc_ct, nrow = 2, labels = NULL)

# Print the final combined plot
print(combined_Vasc_ct)

# Sample data for creating the legend (adjust as necessary)
legend_text <- data.frame(
  x = 1,
  y = 1,
  text = "Diagnosis: Vascular Dementia\nSex: Male\nAge: 75–78\nEducation Years: 10-11\nTotal ACE-III: 74–82"
)

# Create the legend plot with a grey background
legend_plot_Vasc <- ggplot(legend_text, aes(x, y, label = text)) +
  geom_label(size = 5, hjust = 0, fill = "grey90", label.padding = unit(0.5, "lines")) +
  theme_void() +
  theme(plot.margin = margin(0, 0, 10, 0))

# Adjust the width of the plot grid to bring the legend boxes closer
combined_Vasc_ct <- plot_grid(
  legend_plot_Vasc,
  combined_Vasc_ct, 
  ncol = 2,  # Arrange the plots in a single row
  rel_widths = c(2.5, 4)  # Adjust the relative widths if necessary
)

# Add the label "D" to the plot
combined_Vasc_ct <- ggdraw() +
  draw_plot(combined_Vasc_ct) +
  draw_label("D", x = 0.15, y = 0.9, size = 18, fontface = "bold")

# Print the final combined plot with the legend and label
print(combined_Vasc_ct)
```

#Combine AD and FTD maps 

```{r}
combined_AD_FTD <- plot_grid(
  combined_AD_ct, 
  combined_FTD_ct, 
  ncol = 1,  # Arrange the plots in a single column
  rel_heights = c(1, 1)  # Adjust the relative heights if necessary
)

# Add the title to the combined plot
final_AD_FTD <- ggdraw() +
  draw_plot(combined_AD_FTD, 0, -0.03, 1, 1) +  # Slightly move the plot down
  draw_label("Individual Cortical Thickness Outlier Maps", x = 0.5, y = 0.985, size = 16, fontface = "bold", hjust = 0.5)

# Print the final combined plot with the title
print(final_AD_FTD)
```

#Combine Mixed and Vascular maps 

```{r}
# Combine the two plots into a single column layout
combined_Mixed_Vasc <- plot_grid(
  combined_Mixed_ct, 
  combined_Vasc_ct, 
  ncol = 1,  # Arrange the plots in a single column
  rel_heights = c(1, 1)  # Adjust the relative heights if necessary
)

# Print the final combined plot with the title
print(combined_Mixed_Vasc)
```
#----------------------------------------------
# - - - AD Individual Subcortical Volume Maps  - - - 
#----------------------------------------------

#Subject 1 AD 
```{r}
subject1_AD_sv <- df_sv %>% filter(RID == "32651")
subject1_AD_sv <- subject1_AD_sv[,grep("_bin", names(subject1_AD_sv))]
subject1_AD_sv <- gather(subject1_AD_sv)
subject1_AD_sv$ROI <- subject1_AD_sv$key
subject1_AD_sv$ROI <- gsub("_bin", "", subject1_AD_sv$ROI)
subject1_AD_sv <- subject1_AD_sv

# Convert existing data to aseg atlas
data(aseg)
force(aseg)
aseg_labels <- as.data.frame(aseg)

# Merge via labels
subject1_AD_sv <- subject1_AD_sv %>% dplyr::rename("label" = "ROI")
subject1_AD_sv$label <- gsub("\\.", "-", subject1_AD_sv$label)
subject1_AD_sv <- merge(aseg_labels, subject1_AD_sv, by = "label", all.x = TRUE)
subject1_AD_sv <- subject1_AD_sv[grep("region|value|hemi|type", names(subject1_AD_sv))]

# Plot
subject1_plot <- ggseg(.data = subject1_AD_sv, atlas = aseg, mapping = aes(fill = value), colour = "white", size = 0.1) +
  scale_fill_gradientn(limits = c(0.025, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")


# Create a blank plot with the title "Patient 1" at the desired position
title_plot <- ggdraw() + draw_label("Patient 1", x = 0.1, y = 0.5, hjust = 0, size = 14, fontface = "bold") +
  theme(plot.margin = margin(0, 0, 0, 0))

# Create a plot grid with the title positioned above the leftmost brain
subject1_AD_sv <- plot_grid(
  title_plot, 
  subject1_plot,
  ncol = 1,
  rel_heights = c(0.1, 1)  # Adjust the relative heights to position the title
)

# Print the final plot
print(subject1_AD_sv)

```

#Subject 2 AD

```{r}
subject2_AD_sv <- df_sv %>% filter(RID == "36460")
subject2_AD_sv <- subject2_AD_sv[, grep("_bin", names(subject2_AD_sv))]
subject2_AD_sv <- gather(subject2_AD_sv)
subject2_AD_sv$ROI <- subject2_AD_sv$key
subject2_AD_sv$ROI <- gsub("_bin", "", subject2_AD_sv$ROI)
subject2_AD_sv <- subject2_AD_sv

# Convert existing data to aseg atlas
data(aseg)
force(aseg)
aseg_labels <- as.data.frame(aseg)

# Merge via labels
subject2_AD_sv <- subject2_AD_sv %>% dplyr::rename("label" = "ROI")
subject2_AD_sv$label <- gsub("\\.", "-", subject2_AD_sv$label)
subject2_AD_sv <- merge(aseg_labels, subject2_AD_sv, by = "label", all.x = TRUE)
subject2_AD_sv <- subject2_AD_sv[grep("region|value|hemi|type", names(subject2_AD_sv))]

# Plot
subject2_plot <- ggseg(.data = subject2_AD_sv, atlas = aseg, mapping = aes(fill = value), colour = "white", size = 0.1) +
  scale_fill_gradientn(limits = c(0.025, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Create a blank plot with the title "Patient 2" at the desired position
title_plot <- ggdraw() + draw_label("Patient 2", x = 0.1, y = 0.5, hjust = 0, size = 14, fontface = "bold") +
  theme(plot.margin = margin(0, 0, 0, 0)) 

# Create a plot grid with the title positioned above the leftmost brain
subject2_AD_sv <- plot_grid(
  title_plot, 
  subject2_plot,
  ncol = 1,
  rel_heights = c(0.1, 1)  # Adjust the relative heights to position the title
)

# Print the final plot
print(subject2_AD_sv)

```

#Combined SV plots

```{r}
# Combine the two plots
combined_AD_sv <- plot_grid(subject1_AD_sv, subject2_AD_sv, nrow = 2, labels = NULL)

# Print the final combined plot
print(combined_AD_sv)

# Sample data for creating the legend (adjust as necessary)
legend_text <- data.frame(
  x = 1,
  y = 1,
  text = "Diagnosis: Alzheimer's Disease\nSex: Female\nAge: 74–75\nEducation Years: 11-13\nTotal ACE-III: 55–56"
)

# Create the legend plot with a grey background
legend_plot_AD <- ggplot(legend_text, aes(x, y, label = text)) +
  geom_label(size = 5, hjust = 0, fill = "grey90", label.padding = unit(0.5, "lines")) +
  theme_void() +
  theme(plot.margin = margin(0, 0, 10, 0))

# Adjust the width of the plot grid to bring the legend boxes closer
combined_AD_sv <- plot_grid(
  legend_plot_AD,
  combined_AD_sv, 
  ncol = 2,  # Arrange the plots in a single row
  rel_widths = c(2.5, 4)  # Adjust the relative widths if necessary
)

# Add the label "A" to the plot
combined_AD_sv <- ggdraw() +
  draw_plot(combined_AD_sv) +
 draw_label("A", x = 0.15, y = 0.95, size = 18, fontface = "bold")

# Print the final combined plot with the legend and label
print(combined_AD_sv)
```

#----------------------------------------------
# - - - FTD Individual Subcortical Volume Maps  - - - 
#----------------------------------------------

#Subject 1 FTD
```{r}
subject1_FTD_sv <- df_sv %>% filter(RID == "31108")
subject1_FTD_sv <- subject1_FTD_sv[,grep("_bin", names(subject1_FTD_sv))]
subject1_FTD_sv <- gather(subject1_FTD_sv)
subject1_FTD_sv$ROI <- subject1_FTD_sv$key
subject1_FTD_sv$ROI <- gsub("_bin", "", subject1_FTD_sv$ROI)
subject1_FTD_sv <- subject1_FTD_sv

# Convert existing data to aseg atlas
data(aseg)
force(aseg)
aseg_labels <- as.data.frame(aseg)

# Merge via labels
subject1_FTD_sv <- subject1_FTD_sv %>% dplyr::rename("label" = "ROI")
subject1_FTD_sv$label <- gsub("\\.", "-", subject1_FTD_sv$label)
subject1_FTD_sv <- merge(aseg_labels, subject1_FTD_sv, by = "label", all.x = TRUE)
subject1_FTD_sv <- subject1_FTD_sv[grep("region|value|hemi|type", names(subject1_FTD_sv))]

# Plot
subject1_plot <- ggseg(.data = subject1_FTD_sv, atlas = aseg, mapping = aes(fill = value), colour = "white", size = 0.1) +
  scale_fill_gradientn(limits = c(0.025, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Create a blank plot with the title "Patient 1" at the desired position
title_plot <- ggdraw()+
  theme(plot.margin = margin(0, 0, 0, 0))

# Create a blank plot with the title "Patient 1" at the desired position
title_plot <- ggdraw() + draw_label("Patient 1", x = 0.1, y = 0.5, hjust = 0, size = 14, fontface = "bold") +
  theme(plot.margin = margin(0, 0, 0, 0))

# Create a plot grid with the title positioned above the leftmost brain
subject1_FTD_sv <- plot_grid(
  title_plot, 
  subject1_plot,
  ncol = 1,
  rel_heights = c(0.1, 1)  # Adjust the relative heights to position the title
)

# Print the final plot
print(subject1_FTD_sv)

```

#Subject 2 FTD

```{r}
subject2_FTD_sv <- df_sv %>% filter(RID == "33401")
subject2_FTD_sv <- subject2_FTD_sv[,grep("_bin", names(subject2_FTD_sv))]
subject2_FTD_sv <- gather(subject2_FTD_sv)
subject2_FTD_sv$ROI <- subject2_FTD_sv$key
subject2_FTD_sv$ROI <- gsub("_bin", "", subject2_FTD_sv$ROI)
subject2_FTD_sv <- subject2_FTD_sv

# Convert existing data to aseg atlas
data(aseg)
force(aseg)
aseg_labels <- as.data.frame(aseg)

# Merge via labels
subject2_FTD_sv <- subject2_FTD_sv %>% dplyr::rename("label" = "ROI")
subject2_FTD_sv$label <- gsub("\\.", "-", subject2_FTD_sv$label)
subject2_FTD_sv <- merge(aseg_labels, subject2_FTD_sv, by = "label", all.x = TRUE)
subject2_FTD_sv <- subject2_FTD_sv[grep("region|value|hemi|type", names(subject2_FTD_sv))]

# Plot
subject2_plot <- ggseg(.data = subject2_FTD_sv, atlas = aseg, mapping = aes(fill = value), colour = "white", size = 0.1) +
  scale_fill_gradientn(limits = c(0.025, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Create a blank plot with the title "Patient 2" at the desired position
title_plot <- ggdraw() +
  theme(plot.margin = margin(0, 0, 0, 0))

# Create a blank plot with the title "Patient 2" at the desired position
title_plot <- ggdraw() + draw_label("Patient 2", x = 0.1, y = 0.5, hjust = 0, size = 14, fontface = "bold") +
  theme(plot.margin = margin(0, 0, 0, 0))

# Create a plot grid with the title positioned above the leftmost brain
subject2_FTD_sv <- plot_grid(
  title_plot, 
  subject2_plot,
  ncol = 1,
  rel_heights = c(0.1, 1)  # Adjust the relative heights to position the title
)

# Print the final plot
print(subject2_FTD_sv)

```

#Combined SV plots

```{r}
# Combine the two plots
combined_FTD_sv <- plot_grid(subject1_FTD_sv, subject2_FTD_sv, nrow = 2, labels = NULL)

# Print the final combined plot
print(combined_FTD_sv)

# Sample data for creating the legend (adjust as necessary)
legend_text <- data.frame(
  x = 1,
  y = 1,
  text = "Diagnosis: Frontotemporal Dementia\nSex: Male\nAge: 72–75\nEducation Years: 10-13\nTotal ACE-R: 70–76"
)

# Create the legend plot with a grey background
legend_plot_FTD <- ggplot(legend_text, aes(x, y, label = text)) +
  geom_label(size = 5, hjust = 0, fill = "grey90", label.padding = unit(0.5, "lines")) +
  theme_void() +
  theme(plot.margin = margin(0, 0, 10, 0))

# Adjust the width of the plot grid to bring the legend boxes closer
combined_FTD_sv <- plot_grid(
  legend_plot_FTD,
  combined_FTD_sv, 
  ncol = 2,  # Arrange the plots in a single row
  rel_widths = c(3.2, 4)  # Adjust the relative widths if necessary
)

# Add the label "B" to the plot
combined_FTD_sv <- ggdraw() +
  draw_plot(combined_FTD_sv) +
 draw_label("B", x = 0.15, y = 0.95, size = 18, fontface = "bold")

# Print the final combined plot with the legend and label
print(combined_FTD_sv)
```
#-----------------------------------------------
# Mixed Individual Subcortical Volume Maps 
#-----------------------------------------------

#Subject 1 Mixed
```{r}
subject1_Mixed_sv <- df_sv %>% filter(RID == "32934")
subject1_Mixed_sv <- subject1_Mixed_sv[,grep("_bin", names(subject1_Mixed_sv))]
subject1_Mixed_sv <- gather(subject1_Mixed_sv)
subject1_Mixed_sv$ROI <- subject1_Mixed_sv$key
subject1_Mixed_sv$ROI <- gsub("_bin", "", subject1_Mixed_sv$ROI)
subject1_Mixed_sv <- subject1_Mixed_sv

# Convert existing data to aseg atlas
data(aseg)
force(aseg)
aseg_labels <- as.data.frame(aseg)

# Merge via labels
subject1_Mixed_sv <- subject1_Mixed_sv %>% dplyr::rename("label" = "ROI")
subject1_Mixed_sv$label <- gsub("\\.", "-", subject1_Mixed_sv$label)
subject1_Mixed_sv <- merge(aseg_labels, subject1_Mixed_sv, by = "label", all.x = TRUE)
subject1_Mixed_sv <- subject1_Mixed_sv[grep("region|value|hemi|type", names(subject1_Mixed_sv))]

# Plot
subject1_plot <- ggseg(.data = subject1_Mixed_sv, atlas = aseg, mapping = aes(fill = value), colour = "white", size = 0.1) +
  scale_fill_gradientn(limits = c(0.025, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Create a blank plot with the title "Patient 1" at the desired position
title_plot <- ggdraw() + draw_label("Patient 1", x = 0.1, y = 0.5, hjust = 0, size = 14, fontface = "bold") +
  theme(plot.margin = margin(0, 0, 0, 0))

# Create a plot grid with the title positioned above the leftmost brain
subject1_Mixed_sv <- plot_grid(
  title_plot, 
  subject1_plot,
  ncol = 1,
  rel_heights = c(0.1, 1)  # Adjust the relative heights to position the title
)

# Print the final plot
print(subject1_Mixed_sv)
```

#Subject 2 Mixed

```{r}
subject2_Mixed_sv <- df_sv %>% filter(RID == "35015")
subject2_Mixed_sv <- subject2_Mixed_sv[,grep("_bin", names(subject2_Mixed_sv))]
subject2_Mixed_sv <- gather(subject2_Mixed_sv)
subject2_Mixed_sv$ROI <- subject2_Mixed_sv$key
subject2_Mixed_sv$ROI <- gsub("_bin", "", subject2_Mixed_sv$ROI)
subject2_Mixed_sv <- subject2_Mixed_sv

# Convert existing data to aseg atlas
data(aseg)
force(aseg)
aseg_labels <- as.data.frame(aseg)

# Merge via labels
subject2_Mixed_sv <- subject2_Mixed_sv %>% dplyr::rename("label" = "ROI")
subject2_Mixed_sv$label <- gsub("\\.", "-", subject2_Mixed_sv$label)
subject2_Mixed_sv <- merge(aseg_labels, subject2_Mixed_sv, by = "label", all.x = TRUE)
subject2_Mixed_sv <- subject2_Mixed_sv[grep("region|value|hemi|type", names(subject2_Mixed_sv))]

# Plot
subject2_plot <- ggseg(.data = subject2_Mixed_sv, atlas = aseg, mapping = aes(fill = value), colour = "white", size = 0.1) +
  scale_fill_gradientn(limits = c(0.025, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Create a blank plot with the title "Patient 2" at the desired position
title_plot <- ggdraw() + draw_label("Patient 2", x = 0.1, y = 0.5, hjust = 0, size = 14, fontface = "bold") +
  theme(plot.margin = margin(0, 0, 0, 0))

# Create a plot grid with the title positioned above the leftmost brain
subject2_Mixed_sv <- plot_grid(
  title_plot, 
  subject2_plot,
  ncol = 1,
  rel_heights = c(0.1, 1)  # Adjust the relative heights to position the title
)

# Print the final plot
print(subject2_Mixed_sv)

```

#Combined Mixed sv maps

```{r}
# Load necessary libraries
library(ggplot2)
library(cowplot)

# Combine the two plots for Mixed dementia
combined_Mixed_sv <- plot_grid(subject1_Mixed_sv, subject2_Mixed_sv, nrow = 2, labels = NULL)

# Print the final combined plot
print(combined_Mixed_sv)

# Sample data for creating the legend (adjust as necessary)
legend_text <- data.frame(
  x = 1,
  y = 1,
  text = "Diagnosis: Mixed Dementia\nSex: Male\nAge: 80–84\nEducation Years: 11\nTotal ACE-III: 62–78"
)

# Create the legend plot with a grey background
legend_plot_Mixed <- ggplot(legend_text, aes(x, y, label = text)) +
  geom_label(size = 5, hjust = 0, fill = "grey90", label.padding = unit(0.5, "lines")) +
  theme_void() +
  theme(plot.margin = margin(0, 0, 10, 0))

# Adjust the width of the plot grid to bring the legend boxes closer
combined_Mixed_sv <- plot_grid(
  legend_plot_Mixed,
  combined_Mixed_sv, 
  ncol = 2,  # Arrange the plots in a single row
  rel_widths = c(2.5, 4)  # Adjust the relative widths if necessary
)

# Add the label "C" to the plot
combined_Mixed_sv <- ggdraw() +
  draw_plot(combined_Mixed_sv) +
  draw_label("C", x = 0.15, y = 0.95, size = 18, fontface = "bold")

# Print the final combined plot with the legend and label
print(combined_Mixed_sv)

```

#-------------------------------------------------------------------
# - - - Vascular dementia Individual Subcortical Volume Maps  - - - 
#-------------------------------------------------------------------

#Subject 1 Vascular dementia
```{r}
subject1_Vasc_sv <- df_sv %>% filter(RID == "33770")
subject1_Vasc_sv <- subject1_Vasc_sv[,grep("_bin", names(subject1_Vasc_sv))]
subject1_Vasc_sv <- gather(subject1_Vasc_sv)
subject1_Vasc_sv$ROI <- subject1_Vasc_sv$key
subject1_Vasc_sv$ROI <- gsub("_bin", "", subject1_Vasc_sv$ROI)
subject1_Vasc_sv <- subject1_Vasc_sv

# Convert existing data to aseg atlas
data(aseg)
force(aseg)
aseg_labels <- as.data.frame(aseg)

# Merge via labels
subject1_Vasc_sv <- subject1_Vasc_sv %>% dplyr::rename("label" = "ROI")
subject1_Vasc_sv$label <- gsub("\\.", "-", subject1_Vasc_sv$label)
subject1_Vasc_sv <- merge(aseg_labels, subject1_Vasc_sv, by = "label", all.x = TRUE)
subject1_Vasc_sv <- subject1_Vasc_sv[grep("region|value|hemi|type", names(subject1_Vasc_sv))]

# Plot
subject1_plot <- ggseg(.data = subject1_Vasc_sv, atlas = aseg, mapping = aes(fill = value), colour = "white", size = 0.1) +
  scale_fill_gradientn(limits = c(0.025, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Create a blank plot with the title "Patient 1" at the desired position
title_plot <- ggdraw() + draw_label("Patient 1", x = 0.1, y = 0.5, hjust = 0, size = 14, fontface = "bold") +
  theme(plot.margin = margin(0, 0, 0, 0))

# Create a plot grid with the title positioned above the leftmost brain
subject1_Vasc_sv <- plot_grid(
  title_plot, 
  subject1_plot,
  ncol = 1,
  rel_heights = c(0.1, 1)  # Adjust the relative heights to position the title
)

# Print the final plot
print(subject1_Vasc_sv)

```
#Subject 2 Vascular dementia

```{r}
subject2_Vasc_sv <- df_sv %>% filter(RID == "34493")
subject2_Vasc_sv <- subject2_Vasc_sv[,grep("_bin", names(subject2_Vasc_sv))]
subject2_Vasc_sv <- gather(subject2_Vasc_sv)
subject2_Vasc_sv$ROI <- subject2_Vasc_sv$key
subject2_Vasc_sv$ROI <- gsub("_bin", "", subject2_Vasc_sv$ROI)
subject2_Vasc_sv <- subject2_Vasc_sv

# Convert existing data to aseg atlas
data(aseg)
force(aseg)
aseg_labels <- as.data.frame(aseg)

# Merge via labels
subject2_Vasc_sv <- subject2_Vasc_sv %>% dplyr::rename("label" = "ROI")
subject2_Vasc_sv$label <- gsub("\\.", "-", subject2_Vasc_sv$label)
subject2_Vasc_sv <- merge(aseg_labels, subject2_Vasc_sv, by = "label", all.x = TRUE)
subject2_Vasc_sv <- subject2_Vasc_sv[grep("region|value|hemi|type", names(subject2_Vasc_sv))]

# Plot
subject2_plot <- ggseg(.data = subject2_Vasc_sv, atlas = aseg, mapping = aes(fill = value), colour = "white", size = 0.1) +
  scale_fill_gradientn(limits = c(0.025, 1), colours = rainbow.colors(5)) +
  theme(legend.position = "none")

# Create a blank plot with the title "Patient 2" at the desired position
title_plot <- ggdraw() + draw_label("Patient 2", x = 0.1, y = 0.5, hjust = 0, size = 14, fontface = "bold") +
  theme(plot.margin = margin(0, 0, 0, 0))

# Create a plot grid with the title positioned above the leftmost brain
subject2_Vasc_sv <- plot_grid(
  title_plot, 
  subject2_plot,
  ncol = 1,
  rel_heights = c(0.1, 1)  # Adjust the relative heights to position the title
)

# Print the final plot
print(subject2_Vasc_sv)
```

#Combined SV plots 

```{r}
# Combine the two plots for Vascular dementia
combined_Vasc_sv <- plot_grid(subject1_Vasc_sv, subject2_Vasc_sv, nrow = 2, labels = NULL)

# Print the final combined plot
print(combined_Vasc_sv)

# Sample data for creating the legend (adjust as necessary)
legend_text <- data.frame(
  x = 1,
  y = 1,
  text = "Diagnosis: Vascular Dementia\nSex: Male\nAge: 75–78\nEducation Years: 10-11\nTotal ACE-III: 74–82")

# Create the legend plot with a grey background
legend_plot_Vasc <- ggplot(legend_text, aes(x, y, label = text)) +
  geom_label(size = 5, hjust = 0, fill = "grey90", label.padding = unit(0.5, "lines")) +
  theme_void() +
  theme(plot.margin = margin(0, 0, 10, 0))

# Adjust the width of the plot grid to bring the legend boxes closer
combined_Vasc_sv <- plot_grid(
  legend_plot_Vasc,
  combined_Vasc_sv, 
  ncol = 2,  # Arrange the plots in a single row
  rel_widths = c(2.5, 4)  # Adjust the relative widths if necessary
)

# Add the label "D" to the plot
combined_Vasc_sv <- ggdraw() +
  draw_plot(combined_Vasc_sv) +
  draw_label("D", x = 0.15, y = 0.95, size = 18, fontface = "bold")

# Print the final combined plot with the legend and label
print(combined_Vasc_sv)
```

#Combine AD and FTD 

```{r}
# Combine the two plots into a single column layout
combined_AD_FTD_sv <- plot_grid(
  combined_AD_sv, 
  combined_FTD_sv, 
  ncol = 1,  # Arrange the plots in a single column
  rel_heights = c(1, 1)  # Adjust the relative heights if necessary
)

# Add the title to the combined plot
final_AD_FTD_sv <- ggdraw() +
  draw_plot(combined_AD_FTD_sv, 0, -0.03, 1, 1) +  # Slightly move the plot down
  draw_label("Individual Subcortical Volumes Outlier Maps", x = 0.5, y = 0.99, size = 16, fontface = "bold", hjust = 0.5)

# Print the final combined plot with the title
print(final_AD_FTD_sv)

```

#Combine Mixed and Vascular maps - sv 

```{r}
# Combine the two plots into a single column layout
combined_Mixed_Vasc_sv <- plot_grid(
  combined_Mixed_sv, 
  combined_Vasc_sv, 
  ncol = 1,  # Arrange the plots in a single column
  rel_heights = c(1, 1)  # Adjust the relative heights if necessary
)

# Print the final combined plot with the title
print(combined_Mixed_Vasc_sv)

```
